Adding new convention for ports on new rest services
A new WIKI page has been created and referenced in the README.

When creating a new rest project, the first run of ./gradlew bootRun
will cause an exception since server.port and management.port are set to -1.
This was done to solve the problem of multiple backend services on developer
machines from colliding on ports.

https://jira.blackbaud.com/browse/LUM-3383